# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FinFlow API",
    "version": "1.0.0",
    "description": "All endpoints require an authenticated session (Flask-Login cookie, set by /login) and are scoped to the logged-in user's own accounts. State-changing requests (POST/PATCH/DELETE) also require a CSRF token, sent as header `X-CSRFToken` (get one from the `csrf-token` <meta> tag on any page, or the `csrf_token` hidden field on the login/signup forms)."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "parameters": {
      "CsrfHeader": {
        "name": "X-CSRFToken",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Required on POST/PATCH/DELETE."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "integer"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
    "/budgets": {
      "get": {
        "summary": "List per-category budgets with this-month actuals and status",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "month": {
                      "type": "string",
                      "example": "2026-09"
                    },
                    "budgets": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "category_id": {
                            "type": "string"
                          },
                          "category_name": {
                            "type": "string"
                          },
                          "monthly_limit": {
                            "type": "number",
                            "nullable": true
                          },
                          "spent_this_month": {
                            "type": "number"
                          },
                          "projected_month_end": {
                            "type": "number"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "green",
                              "amber",
                              "red",
                              null
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Set (or update) a category's monthly budget limit",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "category_id",
                  "monthly_limit"
                ],
                "properties": {
                  "category_id": {
                    "type": "string"
                  },
                  "monthly_limit": {
                    "type": "number",
                    "minimum": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved"
          },
          "400": {
            "description": "Invalid category or limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/fraud/alerts": {
      "get": {
        "summary": "Flagged transactions with plain-language explanations",
        "parameters": [
          {
            "name": "threshold",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 1
            },
            "description": "Defaults to the user's alert_sensitivity setting."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "threshold": {
                      "type": "number"
                    },
                    "alerts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "transaction_id": {
                            "type": "string"
                          },
                          "merchant_id": {
                            "type": "string"
                          },
                          "amount": {
                            "type": "number"
                          },
                          "date": {
                            "type": "string"
                          },
                          "fraud_probability": {
                            "type": "number"
                          },
                          "is_flagged": {
                            "type": "boolean"
                          },
                          "reasons": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "feedback": {
                            "type": "string",
                            "nullable": true,
                            "enum": [
                              "CONFIRMED_ME",
                              "SUSPICIOUS",
                              null
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "threshold out of [0,1]",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/fraud/feedback": {
      "post": {
        "summary": "Record \"This was me\" / \"Report as suspicious\" on an alert",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "transaction_id",
                  "feedback"
                ],
                "properties": {
                  "transaction_id": {
                    "type": "string"
                  },
                  "feedback": {
                    "type": "string",
                    "enum": [
                      "CONFIRMED_ME",
                      "SUSPICIOUS"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved"
          },
          "400": {
            "description": "Invalid feedback value"
          },
          "404": {
            "description": "Transaction not found (or not owned by this user)"
          }
        }
      }
    },
...
    "/what-if/budget": {
      "post": {
        "summary": "Simulate reducing spend in specific categories",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "add_savings": {
                    "type": "number",
                    "default": 0
                  },
                  "reduce_categories": {
                    "type": "object",
                    "description": "category_id -> reduction fraction (0-1). Keys are whitelisted against dim_category.",
                    "additionalProperties": {
                      "type": "number",
                      "minimum": 0,
                      "maximum": 1
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "monthly_savings_impact": {
                      "type": "number"
                    },
                    "projected_balance_6_months": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Unknown category key(s)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}